# swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "Movie RoR API",
    "description": "Project for Movie",
    "license": {
      "name": "",
      "url": ""
    }
  },
  "host": "localhost:3007",
  "basePath": "/",
...
  "schemes": ["http"],
  "consumes": ["application/json"],
  "produces": ["application/json"],
  "paths": {
...
    "/movie/{id}": {
      "get": {
        "tags": ["Movies"],
        "parameters": [
          { "in": "query", "type": "number", "name": "id", "required": true }
        ],
        "summary": "Get Movies by filters",
        "responses": {
          "200": {
            "description": "Get Movies By Filter",
            "schema": {
              "type": "object",
              "properties": {
                "movie": {
                  "type": "object",
                  "example": {
                      "id": 3,
                      "adult": false,
                      "backdrop_path": "/fQq1FWp1rC89xDrRMuyFJdFUdMd.jpg",
                      "budget": 0,
                      "homepage": "",
                      "imdb_id": "tt13535456",
                      "original_language": "en",
                      "original_title": "Gabriel's Inferno Part III",
                      "overview": "The final part of the film adaption of the erotic romance novel Gabriel's Inferno written by an anonymous Canadian author under the pen name Sylvain Reynard.",
                      "popularity": 18.467,
                      "poster_path": "/fYtHxTxlhzD4QWfEbrC1rypysSD.jpg",
                      "release_date": "2020-11-18T22:00:00.000Z",
                      "revenue": 0,
                      "runtime": 105,
                      "status": "Released",
                      "tagline": "",
                      "title": "Gabriel's Inferno Part III",
                      "genre_ids": [
                        14,
                        4
                      ],
                      "movie_rate": 8.0
                  }
                }
              }
            }
          }
        }
      }
    },
...
  }
}